+++ src/gui/MainWindow.cpp
@@ -3654,14 +3654,18 @@ void MainWindow::wirePanadapter(PanadapterApplet* applet)
             QString("display pan set %1 center=%2").arg(applet->panId()).arg(center, 0, 'f', 6));
     });
     connect(sw, &SpectrumWidget::bandZoomRequested,
-            this, [this, applet]() {
+            this, [this, applet, bandZoomOn = std::make_shared<bool>(false)]() mutable {
+        *bandZoomOn = !*bandZoomOn;
         m_radioModel.sendCommand(
-            QString("display pan set %1 band_zoom=1").arg(applet->panId()));
+            QString("display pan set %1 band_zoom=%2")
+                .arg(applet->panId()).arg(*bandZoomOn ? 1 : 0));
     });
     connect(sw, &SpectrumWidget::segmentZoomRequested,
-            this, [this, applet]() {
+            this, [this, applet, segZoomOn = std::make_shared<bool>(false)]() mutable {
+        *segZoomOn = !*segZoomOn;
         m_radioModel.sendCommand(
-            QString("display pan set %1 segment_zoom=1").arg(applet->panId()));
+            QString("display pan set %1 segment_zoom=%2")
+                .arg(applet->panId()).arg(*segZoomOn ? 1 : 0));
     });
     connect(sw, &SpectrumWidget::filterChangeRequested,
             this, [this](int lo, int hi) {